bin: move index file
Remove bin folder.
