+++ tests/function.js
